+++ src/app/timesheet/timesheet-equipment/timesheet-equipment.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-timesheet-equipment',
+  templateUrl: './timesheet-equipment.component.html',
+  styleUrls: ['./timesheet-equipment.component.css']
+})
+export class TimesheetEquipmentComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}